% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getCHIRPS}
\alias{getCHIRPS}
\title{Function to download CHIRPS data}
\usage{
getCHIRPS(
  tres = "monthly",
  sres = 0.05,
  dates = NULL,
  dsn = getwd(),
  overwrite = FALSE
)
}
\arguments{
\item{tres}{numeric. Temporal resolution. \code{"daily"} or \code{"monthly"}}

\item{sres}{numeric. Spatial resolution. \code{0.05} or \code{0.25}. Note
that monthly data is only available at a resolution of 0.05 degrees.}

\item{dates}{date. Date(s) for which you want to download CHIRPS data}

\item{dsn}{path. Directory to which you want to store the downloaded data}

\item{overwrite}{logical. In case the file already exists, should it be
overwritten?}
}
\value{
\code{RasterStack}
}
\description{
Function to download CHIRPS raster data
}
\examples{
\dontrun{
# Load raster package
library(raster)

# Download data
files <- getCHIRPS(
    tres       = "daily"
  , sres       = 0.25
  , dates      = as.Date("2016-12-31")
  , dsn        = tempdir()
  , overwrite  = T
)

# Load the files as a rasterstack
rain <- stack(files)

# Reclassify values (get rid of -9999)
rain <- reclassify(rain, c(-Inf, 0, NA))

# Visualize
plot(rain)
}
}
